Moved actor classes into "actors" package
