+++ client/src/main/scala/it/cwmp/client/controller/actors/GameViewActor.scala
@@ -1,9 +1,10 @@
-package it.cwmp.client.controller
+package it.cwmp.client.controller.actors
 
 import java.util.concurrent.ThreadLocalRandom
 
 import akka.actor.{Actor, ActorRef, Cancellable}
-import it.cwmp.client.controller.GameViewActor._
+import it.cwmp.client.controller.actors.GameViewActor._
+import it.cwmp.client.controller.actors.common.{ActorAlertManagement, AlertMessages}
 import it.cwmp.client.controller.game.GameConstants.{MAX_TIME_BETWEEN_CLIENT_SYNCHRONIZATION, MIN_TIME_BETWEEN_CLIENT_SYNCHRONIZATION}
 import it.cwmp.client.controller.game.GameEngine
 import it.cwmp.client.controller.messages.Initialize
