+++ client/src/main/scala/it/cwmp/client/controller/actors/PlayerActor.scala
@@ -1,11 +1,11 @@
-package it.cwmp.client.controller
+package it.cwmp.client.controller.actors
 
 import akka.actor.{Actor, ActorRef, AddressFromURIString, Props, Stash}
 import akka.cluster.Cluster
 import akka.cluster.ClusterEvent._
 import akka.cluster.ddata.DistributedData
-import it.cwmp.client.controller.GameViewActor.{NewWorld, ShowGUIWithName}
-import it.cwmp.client.controller.PlayerActor.{PrepareForGame, RetrieveAddress, RetrieveAddressResponse}
+import it.cwmp.client.controller.actors.GameViewActor.{NewWorld, ShowGUIWithName}
+import it.cwmp.client.controller.actors.PlayerActor.{PrepareForGame, RetrieveAddress, RetrieveAddressResponse}
 import it.cwmp.client.controller.game.generation.GenerationStrategy
 import it.cwmp.client.controller.messages.{Initialize, Request, Response}
 import it.cwmp.client.model.game.distributed.AkkaDistributedState
