+++ client/src/main/scala/it/cwmp/client/controller/actors/common/ActorAlertManagement.scala
@@ -1,7 +1,7 @@
-package it.cwmp.client.controller
+package it.cwmp.client.controller.actors.common
 
 import akka.actor.Actor.Receive
-import it.cwmp.client.controller.AlertMessages._
+import it.cwmp.client.controller.actors.common.AlertMessages._
 import it.cwmp.client.view.FXAlertsController
 
 /**
