+++ client/src/main/scala/it/cwmp/client/view/FXServiceViewActor.scala
@@ -1,8 +1,8 @@
 package it.cwmp.client.view
 
 import akka.actor.{Actor, ActorRef}
+import it.cwmp.client.controller.actors.common.{ActorAlertManagement, ActorViewVisibilityManagement}
 import it.cwmp.client.controller.messages.Initialize
-import it.cwmp.client.controller.{ActorAlertManagement, ActorViewVisibilityManagement}
 
 /**
   * A base class representing a Service View actor with JavaFX underlying
